refactor: 更新
Former-commit-id: 45fff08eb868830e6c27f9c532f763b6e0051697

+++ core/model/src/main/java/com/imcys/model/download/Entry.kt
@@ -37,7 +37,7 @@ data class Entry(
     @SerialName("owner_avatar")
     val ownerAvatar: String = "",
     @SerialName("owner_id")
-    val ownerId: Int = 0,
+    val ownerId: Long = 0,
     @SerialName("owner_name")
     val ownerName: String = "",
     @SerialName("page_data")
